Fix playable video URL pattern
The previous pattern assumes that the JSON blob in the embed video page
always comes back with attributes in the same order, but this is not the
case

Instead it now searches within each video quality array for anything
that's not the end of the array followed by a "url" attribute. It also
uses the `*?` [1] patterns for non-greedy matches, thus simplifying the
regexp as a whole.

[1]: https://docs.python.org/2/howto/regex.html#greedy-versus-non-greedy