Fix tests that broke due to change in website
This test was already broken before the recent changes due to changes in
the API response.  As it is, it's still brittle since every time the
response changes, it could potentially break.

Going forward, if it keeps breaking, we should probably mock the API
response and maybe add contract tests to verify that the website
responds as expected.